Conversation Editor and Questions Components
Added the stakeHolder Component to the conversation editor component page. Also added the questions component to the reflection page. Components created are in their own folders so there is no confusion of what goes where. Didnt delete any of the previous conversation editor parts, just commented the lines out. Still working on styling for mobile use, havent looked into it yet, but can be on the shelf atm.